+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-07-30 20:18:36
-// Commit: fix(perf): fix memory leak by cancelling perf subscription
+// Updated at: 2026-08-02 17:24:51
+// Commit: fix(perf): prevent double tap duplicate action in perf
 
 class perfComponentState {
-  final String commitRef = '3261';
-  final String lastMessage = 'fix(perf): fix memory leak by cancelling perf subscription';
-  final DateTime timestamp = DateTime.parse('2026-07-30T20:18:36+05:00');
+  final String commitRef = '3280';
+  final String lastMessage = 'fix(perf): prevent double tap duplicate action in perf';
+  final DateTime timestamp = DateTime.parse('2026-08-02T17:24:51+05:00');
 }